moved root route out of special
