+++ src/plugins/special/special.ts
@@ -1,5 +1,5 @@
 import type { FastifyInstance, FastifyPluginOptions, FastifyRequest, FastifyReply } from 'fastify';
-import logger from '../../utils/logger.js';
+import logger from '../../../utils/logger.js';
 
 /**
  * @param {FastifyInstance} fastify
@@ -7,22 +7,7 @@ import logger from '../../utils/logger.js';
  */
 async function specialRoutes(fastify: FastifyInstance, options: FastifyPluginOptions) {
 
-    fastify.get('/', async (request) => {
-      return Welcome(request);
-    });
-
     fastify.get('/user', UserDetails);
-}
-
-// This function is independent of Fastify's route registration
-export function Welcome(request: FastifyRequest) {
-  logger.info(
-    `Incoming ${request.method} request from ${request.ip} to ${request.headers.host}${request.url}`
-  );
-
-  return {
-    message: 'Hello and welcome to the UK Location API!',
-  };
 }
 
 /**
@@ -49,5 +34,6 @@ function UserDetails(ctx) {
     user,
   };
 }
+
 
 export default specialRoutes;
